Backend refactor and new fields for nodes added

+++ backend/src/main/java/com/kiwi/features/nodes/controllers/UserNodeStatusRepository.java
@@ -7,5 +7,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 public interface UserNodeStatusRepository extends JpaRepository<UserNodeStatusPersistence, UserNodeStatusKey> {
-    Optional<UserNodeStatusPersistence> findByIdUserIdAndIdNodeId(Long userId, int nodeId);
+    Optional<UserNodeStatusPersistence> findByIdUserIdAndIdNodeId(Long userId, Long nodeId);
 }
