+++ backend/src/main/java/com/kiwi/features/nodes/exceptions/NodeLockedException.java
@@ -3,7 +3,7 @@ package com.kiwi.features.nodes.exceptions;
 import jakarta.validation.constraints.NotNull;
 
 public class NodeLockedException extends RuntimeException {
-    public NodeLockedException(@NotNull int id) {
+    public NodeLockedException(@NotNull Long id) {
         super(String.format("Cannot unlock node with id %s", id));
     }
 }
