+++ backend/src/main/java/com/kiwi/features/nodes/exceptions/NodeMarkAsLockedException.java
@@ -3,7 +3,7 @@ package com.kiwi.features.nodes.exceptions;
 import jakarta.validation.constraints.NotNull;
 
 public class NodeMarkAsLockedException extends RuntimeException {
-    public NodeMarkAsLockedException(@NotNull int id) {
+    public NodeMarkAsLockedException(@NotNull Long id) {
         super(String.format("Cannot mark next nodes as locked because node with id %s is not COMPLETED", id));
     }
 }
