+++ backend/src/main/java/com/kiwi/features/nodes/exceptions/NodeSatusNotFoundException.java
@@ -3,7 +3,7 @@ package com.kiwi.features.nodes.exceptions;
 import jakarta.validation.constraints.NotNull;
 
 public class NodeSatusNotFoundException extends RuntimeException {
-    public NodeSatusNotFoundException(@NotNull int id) {
+    public NodeSatusNotFoundException(@NotNull Long id) {
         super(String.format("Cannot found status of node with id %s", id));
     }
 }
